+++ drivers/uart/uart.c
@@ -7,6 +7,7 @@
 
 #include "drivers/uart/uart.h"
 #include "hal/uart/uart_irda_cir.h"
+#include "hal/uart/uart_console.h"
 
 #include "hw_uart_irda_cir.h"
 #include "soc_AM335x.h"
@@ -14,57 +15,12 @@
 /* Utils */
 #include "utils/string.h"
 
-/* System config */
-#include "kernel/config.h"
-
-/*****************************************************************************
-**                      INTERNAL MACRO DEFINITIONS
-*****************************************************************************/
-#define BASE_10             (10u)
-#define BASE_16             (16u)
-#define INVALID_INPUT       (0xFFu)
-#define MAX_STRING_WIDTH    (80u)
-
-/*
-** Loop continuously until user enters any character other
-** than space, carriage return, tab space and backspace
-*/
-#define IS_WHITESPACE(rxByte)    rxByte = UARTGetc(); \
-                                 while(('\r' == rxByte) || (' ' == rxByte) || \
-                                       ('\t' == rxByte) || ('\b' == rxByte)) { \
-                                      UARTPutc(rxByte);   \
-                                      rxByte = UARTGetc(); }
-
-/* 
-** Echoes back the delimiting character entered by the user
-** after the actual input value has been read from the user.
-*/
-#define UART_SCANF_ECHO_INPUT(rxByte)                                     \
-                                 if(('\n' == rxByte) || ('\r' == rxByte)) \
-                                 {                                        \
-                                     UARTPutc('\r');                      \
-                                     UARTPutc('\n');                      \
-                                 }                                        \
-                                 else                                     \
-                                     UARTPutc(rxByte);
-
-/* A mapping from an integer between 0 and 15 to its ASCII character
- * equivalent. */
-static const char * const g_pcHex = "0123456789abcdef";
-
-/****************************************************************************
-**                    EXTERNAL FUNCTION DECLARATIONS
-****************************************************************************/
-extern void UARTConsolePutc(unsigned char data);
-extern unsigned char UARTConsoleGetc(void);
-extern void UARTConsoleInit(void);
-
 void UART_StdioInit(void)
 {
     UARTConsoleInit();
 }
 
-static inline void UART_stdioRead(unsigned char *rxBuff, unsigned char rxByte)
+void UART_stdioRead(unsigned char *rxBuff, unsigned char rxByte)
 {
     unsigned int inputCount = 0u;
 
@@ -186,12 +142,3 @@ unsigned int UART_writeStr(const char *str) {
 
     return written;
 }
-
-
-
-
-
-
-
-
-
