+++ kernel/makesrc
@@ -1,3 +1,3 @@
+interrupt.o
 boot.o
 kernel.c
-drivers/drivers.c
